Rename libtracker-common to libtracker-miners-common
We made a big compromise when splitting tracker core from tracker-miners
in that the common code that was needed by both parts would end up
duplicated. It's ugly but it works fine at the moment and allows us to
keep all of the common code private.

I had an issue when trying to embed tracker core into tracker-miners as
a Meson subproject though. Having two targets named tracker-common
caused confusion as duplicate targets aren't allowed, but they are not
quite equivalent so we can't just pick one or the other.

To work around this, I've renamed the copy in this repo to
tracker-miners-common. This only affects the target names, not the
actual function names.

https://bugzilla.gnome.org/show_bug.cgi?id=789725