tracker-extract: Delegate activation on tracker-miner-fs
This process is safe to shutdown on inactivity, so let's just do
that. tracker-miner-fs was already partially in charge of
tracker-extract's lifetime through its watchdog. With this commit
it is now totally in charge of it, autostarting it whenever there's
any content that hasn't gone yet through it.

In order to keep recovering from crashes from tracker extract,
perform the check whenever the DBus name is lost, it'll be a no-op
if tracker-extract is legitimately shutting down due to inactivity.