Allow passing a full path to a domain rule
This extends the TrackerDomainOntology class to handle domain rules
which are not installed into Tracker's configured prefix. The domain
name is now interpreted as a full path if the first character is `/`.

Some code was using the input domain rule to determine the name of the
domain. This breaks when the input domain rule is a full path. The
domain rule contains the domain name, so this isn't a problem but in
order to make things work I had to change the behaviour of
tracker_domain_ontology_get_domain() as previously it would append
".Tracker1" onto whatever the domain rule told us the name of the
domain was. Now it does not do that, so it can be used to get the
name of the domain itself as well as the names that the Tracker daemons
running inside that domain should use.

The motivation for this change is that it allows unprivileged users to
make use of the domain ontologies feature. Previously it could only be
used by those who had permissions to install the .rule file into /usr.

Some extra work is still needed to make the feature easy to use for
unpriviliged users, as there also need to be D-Bus service files
created for the domain. This can be done as an unpriviliged user by
starting a separate D-Bus instance and looks for .service files in
a different place, but it's not ideal to require that effort.