functional-tests: Fix more issues in awaiting for property changes
We actually get a GraphUpdated signal per RDF class, and we need to
filter out those we don't care about.

This was breaking tests in some cases as they would get double
notifications and get ahead of themselves. For example the
410-extractor-decorator test would wait for changes to nie:title on
an nmm:MusicPiece resource. But music files also have class nfo:Audio
and so there would be two GraphUpdated signals for each change, one
for the nmm:MusicPiece class and one for the nfo:Audio class.