tracker-extract-gstreamer: Fix iso 8601 string generation in negative TZs
We try to format the timezone hour difference with %02d, however for
negative numbers it will contain just eg. "-4", which is not the expected
behavior.

So, take the negative sign out of the integer, and add it separately
just like we do for the explicit '+'. This makes the %02d bit actually
correct, so the timezone bit starts with "-04".

https://gitlab.gnome.org/GNOME/tracker/issues/22

Closes: tracker#22