libtracker-extract: Use only date bits in album/disc URN
The full date+time will be more prone to inexactitudes, as it's not likely
to be the actual release time of the album (not to the *second*, anyway).
If minutes/seconds are found, it's more likely creation/ripping time.

In that case, the tag still could be based on wrong metadata, but songs
won't be split in multiple albums.