Allow use of domain rules that aren't installed into /usr
Currently Tracker domain rules must be installed inside Tracker's data
directory (usually /usr/share/tracker). This is limiting as it means
only system packages can add them. A program installed into /opt is
unable to use a custom domain, for example. Since Tracker is implemented
as a system of daemons, it's not particularly straight forward to work
around this by setting XDG_DATA_DIRS= to point somewhere non-standard
either.

This patch removes this restriction in a simple way: it allows users to
pass a full path to the domain rule, rather than just the base name.