Ignore some more files that can appear in the applications directory
Since commit d13fd8b28a7c95a1f1bced we have been mining .desktop files
from the XDG application directories such as /usr/share/applications. In
order for this to work as expected, we need the FS miner to ignore other
files that it finds there. This commit adds 2 files that appear on my
Fedora system to the ignored-files list.

If there are files in these directores that aren't expected to be there,
it breaks the functional-tests. The 300-miner-basic-ops test creates 3
text files and asserts that these and only these were found by the
miner-fs. If extra stuff from XDG applications dirs turns up in the list
of text files, the test fails.