functional-tests: Forbid GStreamer to fork for the registry
Running functional tests in an isolated environment result on gst
trying to build the registry again, and fork for it. This results
in a g_spawn call, and who knows why, in this debug line written to
stdout:

posix_spawn avoided (fd close requested)

This is a g_debug() call from glib which should not end up on screen
with our verbosity settings, so I'll asume someone else is playing
smart and replacing the default log handler with something that
introduces garbage on our stdout.

This notably causes JSON output in video tests to be not quite JSON,
causing test failures.