tracker-extract: Relate contained audio tracks and container file both ways
The migration to TrackerResource broke the handling of flac/cue, as
tracker_resource_print_*() do expect all related resources to be directly
referenced from the root resource, but the gstreamer extractor just added
backreferences from the nmm:MusicPieces to the parent nfo:FileDataObject.

Add a relation both ways through nie:hasLogicalPart/nie:isLogicalPartOf
so the graph is complete, and the tracks referenced from the root element.

Closes: https://gitlab.gnome.org/GNOME/tracker-miners/issues/60