tracker-extract: Only set nfo:Audio on rules handling FLAC files
This is the most generic class that still triggers GraphUpdated, we need
to delegate on the extractor module to figure out whether the file is a
music piece on its own or a container of music pieces.

In the latter case the role belongs on the contained elements, not on the
file resource itself, so avoid setting it here for all.