tracker-extract-text: Correctly report errors back to the caller
The tracker_extract_get_metadata() function should return FALSE if
an error occurs reading the file. The tracker-extract-text module would
return TRUE in all cases.

This was causing intermittent failures in the functional-tests, as the
following situation could occur:

  1. file2.txt is created
  2. tracker-miner-fs sees file2.txt and processes it
  3. file2.txt is deleted (or moved into an unmonitored directory)
  4. tracker-miner-fs sees the deletion and removes its resource from
     the store
  5. tracker-extract sees the created notification for file2.txt and
     tries to process it
  6. the tracker_extract_get_metadata() incorrectly returns TRUE
     (success), so tracker-extract recreates the deleted resource

This problem was being detected in the functional tests and was
causing intermittent failures.

This hopefully fixes: https://gitlab.gnome.org/GNOME/tracker-miners/issues/56