libtracker-data: Pass virtual table data on TrackerDBManager construction
There are paths where the TrackerDBManager may create the writable interface
early within tracker_db_manager_new(), setting the vtab data later will
work for database interfaces created later on, but not for the already
created ones.

Pass the vtab data as (yet...) another argument to tracker_db_manager_new()
so it is available anytime for the interfaces that will be later created.

Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=1683903