tracker-extract-gstreamer: Extract MusicBrainz IDs
Extraction of MusicBrainz IDs is done based on the tags defined by
Gstreamer.

Some tags are already part of the stable GStreamer releases:
 - recording id (GST_TAG_MUSICBRAINZ_TRACKID)
 - release id (GST_TAG_MUSICBRAINZ_ALBUMID)
 - artist id (GST_TAG_MUSICBRAINZ_ARTISTID)

Some tags are not part of a GStreamer release yet (they will be
introduced in the 1.18.0 version):
 - track id (GST_TAG_MUSICBRAINZ_RELEASETRACKID)
 - fingerprint id (GST_TAG_ACOUSTID_FINGERPRINT)
 - release-group id (GST_TAG_MUSICBRAINZ_RELEASEGROUPID)
See:
https://gitlab.freedesktop.org/gstreamer/gst-plugins-base/-/merge_requests/252

In the later case, the defined operator is used to check if the tag is
supported.

The flac extraction functional test is updated to handle recording id,
release id and artist id.

Closes: https://gitlab.gnome.org/GNOME/tracker-miners/issues/66