build: Remove old extract rules from build tree when configuration changes
For the functional-tests we run Tracker entirely from the source tree.
We need to maintain a directory of .rule files for tracker-extract to
use, and we must only have the .rule files that are enabled in the
current build configuration. This wasn't working properly because we
didn't remove old .rule files when reconfiguring, so disabling an
extractor at configure-time would sometimes not work as expected.

This is fixed by preparing the uninstalled rule files in a shell
script rather than from Meson, allowing us to delete the rule files
each time that we reconfigure.