Don't recurse into subdirectories when indexing ~/Downloads
We can't really make assumptions about the content of folders in
the 'Downloads' directory. People may download any kind of content
from the internet and then extract it "in place" in the Downloads
directory.

This change was triggered by
https://gitlab.gnome.org/GNOME/tracker/issues/95 where several
users reported system lockups caused by Tracker scanning video
game data. The goal of this change is not to avoid the lockups,
but to avoid scanning game data directories in the first place.
Games often contain lots of audio, video, images and text files,
which can end up in unwanted places like Gnome Photos and Gnome
Music if we do scan it with Tracker.

This commit changes the default settings so that we index files in
~/Downloads but not directories. Hopefully this gives the right
balance between: providing search results for files the user
just downloaded, but not mixing up video game data with actual user
content, and not consuming lots of resources unnecessarily.