functional-tests: Rework the writeback test case baseclass
This is in preparation for adding audio writeback test cases.

As well as removing some assumptions about the test data, we now use
the 'pathlib' module to handle paths. In particular we can properly
generate URIs from filenames now using the .as_uri() method. This
does mean we now require Python 3.4 at minimum to run the functional
tests, but that's a 5 year old release already.