build: Compile GSettings schemas after install
This isn't needed in the common case of distro package builds, as the
distro package will already provide a hook. However, our README.md
tells developers to install into a prefix like ~/opt/tracker, and
that is broken by default because GLib will complain that none of the
settings schemas are installed.