tracker-extract-gstreamer.c: Set empty tags list if no tags were found
If entry doesn't has tags then `tracker_toc_add_entry` receives NULL
instead of the tag list. But this function immediately tries to make a
reference to the tag list and fails on NULL pointer. And every function
in `tracker-cue-sheet.c` assumes that the list is non-NULL.

Now it creates empty tag list if an entry doesn't contain tags. It
shouldn't have any impact since in normal case tag list is copied.