Replace TrackerDomainOntology class with a simpler function
This allows us to remove some code from libtracker-miners-common, but
the motivation is actually to fix a name conflict between
libtracker-common and libtracker-miners-common.

Previously we were prevented from linking statically to
libtracker-miner, and prevented from using link-time optimization,
due to the duplicate GTypes.

The code is now diverged from that in libtracker-common, but it's
unlikely to see major changes because the "domain ontology" format is
part of Tracker's public API.

Fixes: https://gitlab.gnome.org/GNOME/tracker/issues/116