functional-tests: Check for IPCT image tags only if libiptcdata is available
test-image-1 was checking for the presency of image tags that are added only
when libiptcdata is found.

However the test case checks for them in all the cases, so split the image-1
test in two extractor tests, where one is doing the standard checks and
another one is only checking the tags provided by libiptcdata.

Also move images/test-iptcdata-records check there