ci: Run a maximum of 4 tests simultaneously
This is an attempt to deal with the random test failures that sometimes
appear.

The CI runners on gitlab.gnome.org report 32 CPU cores, so Meson tries
to run 32 tests at a time. I guess because our tests are mostly IO
bound this actually makes things slower and causes random failures.