tracker-extract-mp3: Improve duration computation for VBR files
When a file is encoded with variable bitrate mode (VBR), the bitrate
of each frame can be different. In that case, It is necessary to know
the average bitrate in order to compute the duration of the file. This
calculation requires to loop through all the frames. This operation
being too slow, Tracker loops only through the frames up to a
threshold (MAX_FRAMES_SCAN) to approximate the average bitrate. This
approximation can result in a computed duration being very different
from the real duration.

However, most of VBR files also have a XING header. This header can
contain the total number of frames. By using the number of frames, it
is possible to correctly compute the duration with the following
formula:

duration = number_of_frames * samples_per_frame / sampling_rate

Closes: #88