Use the actual IDs for Musicbrainz resources
We were generating internal unique IDs for Musicbrainz resources,
which looked like this:

    urn:ExternalReference:https:%2F%2Fmusicbrainz.org%2Fdoc%2FArtist:76b2e842-5e85-4c97-ab62-d5bc315595b5

The actual Musicbrainz URL is already guaranteed to uniquely represent
the resource, so let's use that to identify the resource inside Tracker.
This means the URL is shorter, more readable, and can be browsed. The
above example becomes this:

    https://musicbrainz.org/artist/76b2e842-5e85-4c97-ab62-d5bc315595b5

This shouldn't be backported to the stable branch.