Use pthreads API to set SCHED_IDLE scheduling policy
Previously we were using the GLIBC sched_setscheduler() function.
This is Linux-only.  We now use the pthread_setschedparam() function
from pthreads, which is portable beyond Linux and works with
musl libc as well as GLIBC.

While the documentation for sched_setscheduler() claims that it affects
the whole process, in fact it only affects the calling thread - see
https://www.openwall.com/lists/musl/2016/03/01/5 for details