tracker-extract: Process small / empty text files
Previously small or empty text files were ignored by the extractor.
This could cause inconsistent results in the database, for example
if the user truncated a regular text file to 0 bytes then the old
content would remain in the store.

We now always update the nie:plainTextContent field for small and
0-byte files.