tracker-extract: Set min glib version where we use libgsf
Libgsf has API that uses GParameter, a struct marked as deprecated
in glib 2.54. As it seems unlikely it'll shed that API any soon,
make it sure complation doesn't trigger deprecation warnings we
don't fall in ourselves.

We define it in tracker-gsf and tracker-extract-msoffice itself,
constrained to C files so we ensure the define doesn't propagate
unintendedly.