tracker-miner-fs: Do not handle writeback file updates specially
Currently, tracker-miner-fs expects an "updated" inotify event after
returning from a PerformWriteback call, this typically meaning that
tracker-writeback did it's job.

This however may fail in 2 ways:
- The related inotify event arrives before the PerformWriteback
  callback is called. If this happens, tracker-miner-fs will ignore
  it and await for an extra one that might never arrive. This was
  probably made much more likely to happen after commit 3bc295d2410,
  as all inotify events had a pretty high timeout.
- The directory is not monitored at all, this may happen due to
  various reasons, like settings, or inotify handle/client limits.

The outcome then is that the miner is left paused waiting for inotify
events that may never arrive.

Since it's not possible to ignore "writeback updates" in a non-racy
manner, give it up and handle the file update as every other.

Closes: https://gitlab.gnome.org/GNOME/gnome-music/issues/346