miners/fs: Remove all data added by IndexFileForProcess on shutdown
The IndexFileForProcess method is intended for apps to trigger on-demand
indexing of removable devices. The data remains in the store as long as
the app is connected to D-Bus, then it's deleted.

If the miner-fs shut down before the app, the data would remain in the
store forever and would not be updated, until an app one day called
IndexFileForProcess again for the same location.