Don't start the daemon for the root user
Now that the .service files are started by the default.target, the
processes could also be started when a cronjob is starting, this is
probably not wanted.

In addition to that, running the extractor as root could lead to
security issues. Let's do like pulseaudio and not start the daemons for
the root user.