tracker-extract: Set a deadline of 30 seconds for extraction tasks
If we stick longer than that for a single file, it would seem there's
some trouble processing it. As we're no longer guaranteed to stay in
control of the extraction thread, just exit the process. As the file
was being processed at the time of exiting, it would fall through the
retry/ignore paths we have in place for extractor crashes.

This timeout executes on the same thread/main context that called
tracker_extract_file(). It is guaranteed to trigger since we offload
extraction tasks to other threads.

This makes tracker-extract resilient to possible infinite loops that
might happen in 3rd party code via extractor modules.

Fixes: https://gitlab.gnome.org/GNOME/tracker-miners/-/issues/107