tracker-extract-gstreamer: Distinguish artist and performer tags
Since tracker!261, there is a difference in the ontology between a
'nmm:artist' (the main artist of a song) and a 'nmm:performer' (a
musician that plays on the song).

The gstreamer extractor tries to use the performer tag
first and uses the artist tag if the performer is not present. The
result is mapped to nmm:performer.

This commit introduces a new behavior to reflect the ontology change:
- 'GST_TAG_PERFORMER' maps to 'nmm:performer'
- 'GST_TAG_PERFORMER' maps to 'nmm:artist'