tracker-extract: Take over all nie:InformationElement data
Tracker-miner-fs will not fill nie:InformationElement for us, it
will just insert a nfo:FileDataObject in the relevant graphs (other
than tracker:FileSystem) and it is our duty to inspect those files.

Luckily, we already inspect the file (and its mimetype) directly
using GIO, so everything is in place for tracker-extract to handle
completely with nie:InformationElement.

Also, prefer the tracker:extractorHash property as a mean to tag
files, as this will allow to track changes in extractors.