libtracker-miner: Adapt TrackerDecorator to nie:InformationElement changes
We must not track nie:InformationElements anymore, but nfo:FileDataObject.
The query for missing elements has been greatly simplified, we now just
track nfo:FileDataObjects in the different graphs that do not have a
tracker:extractorHash in the tracker:FileSystem graph. Using offset in
the query instead of NOT IN for the current/blacklisted elements makes
it possible to reuse the query.

This does take big shortcuts, that will likely turn into further future
cleanups.