tracker-miner-fs: Drop query for unextracted items
Simply keep track of whether we did process any file, and trigger
extraction in that case. This misses the case where tracker-extract
crashes and we used to query again before re-starting, however we'll
rely on systemd for that as everything is in place to handle abnormal
exits.

This makes the tracker-extract watchdog largely ineffective too.