tests: Skip libtracker-miner/tracker-miner-fs-test
This test exercises libtracker-miner by implementing a mini-tracker-miner-fs.
It made some sense when libtracker-miner was public API, now it is not
public anymore, and pretty deep behavioral changes are being done up and
down.

This requires this mini implementation to catch up with the real
tracker-miner-fs, but at this point it is worth wondering how fictional is
this test.

We have functional tests exercising the real thing, we should check that
these functional tests don't miss any corner we cover here, but in the
mean time simply skip these tests.