tracker-control: Move files_indexed/graph properties to a separate interface
This is not the "public API" that we intend applications to use, better
to keep the XML definition for it as clean as possible.

Technically this object/interface is still available to clients, if this
DBus name were made available on a sandbox. And technically letting
sandboxed clients potentially deduce IndexFile requests from other
clients is an information leak, it is just not one to care hard about.

These properties could presumably turn into targeted signals to
tracker-miner-fs/tracker-extract DBus names, and DBus name checks on
respective DBus requests to ensure no one else has access to these
properties. Let's go for simplicity at the moment, it's an implementation
detail and may change anytime.