tests: Drop assert requiring perfect multivalued resource matches
The extractor tests were mandating that the extracted resource
perfectly match the reference json file. However this brings problems
now that we have tags that depend on the gstreamer (plugins) version.

So, don't check as much thorouhgly atm, we'll think later on about
how to check the missing acoustid/mb tags with varying gstreamer
versions.