tracker-miner-fs: Drop special case for update vs create
We query the URN here as a mean to check that a file was already
added, and do things that should essentially turn into no-ops in
the case of a create event (as the resource would exist previously).

Drop the URN check here, makes create/update events handled all the
same, and avoids having to query information elements for things
that are not folders.

Besides, this fixes a funky side effect of our event coalescing wrt
writeback: We write the file in a temporary hidden file and then
move onto the original location. Sometimes we get to coalesce
CREATE(A)+MOVE(A,B) as CREATE(B) right away, without checking that
B exists. This skips the the extractorHash property deletion, so
tracker-extract doesn't get to re-extract the file. If we handle
creates and updates the same way, this is moot.