libtracker-miner: Do not force URN query from the upper layers
With the pre-caching now done correctly for all files we care about
(folders basically), we can avoid forcing queries here if the URN is
not found.

Results in less queries on first-time index, since we are querying
anyway for URNs that we know are not there.

It is worth noting that queries will be implicitly forced here for
folders with an invalidated URN, this only happens if the queried
file was freshly inserted in the DB.