tracker-miner-fs: Process only one item at a time
Despite us spawning a number of tasks (10 currently) to process
files in parallel, disk access is a bottleneck to them. Reduce
this to 1, it doesn't seem to impact performance that much, and
it does help in not processing files together that may not go
in the same SPARQL update batch.

This is irrelevant now, since the operation would be blocked by
the parent being still processed, but we want to lighten that
block in future commits.