tests: Add a performer tag to the vorbis audio file
Since tracker!261, there is a difference in the ontology between a
'nmm:artist' (the main artist of a song) and a 'nmm:performer' (a
musician that plays on the song).

For the moment, the gstreamer extractor tries to use the performer tag
first and uses the artist tag if the performer is not present. The
result is saved in 'nmm:performer'. That's why, adding a performer tag
to the test files changes the result of the tests.
This behavior is now incorrect. A follow-up will introduce a
distinction between artist and performer.