tracker-extract-mp3: Change performer fields to artist
Since tracker!261, there is a difference in the ontology between a
'nmm:artist' (the main artist of a song) and a 'nmm:performer' (a
musician that plays on the song).

For the moment, the mp3 extractor correctly extracts the artist
information and saves it in 'nmm:performer'. The artist information is
now is saved in 'nmm:artist' to reflect the ontology change.

The different variables are renamed from performer to artist to avoid
any confusion.