libtracker-miner: Fix accounting of TrackerFileSystem
After dealing with changes and emitting events on files, the
TrackerFileNotifier intends to only keep the directory skeleton in the
TrackerFileSystem, and forget about regular files (until they might be
needed again).

The code here fared well when we used to query and crawl folders one
by one, as it was meant to run once per folder. We however changed to
querying per index-root and crawling per folder, so this is now forgetting
about files in a random directory of the index root and leaving everything
else.

Forget about the files from the root directory, this restores the neat
barebones TrackerFileSystem structure after indexing.