tracker-miner-fs: Make graph-specific data optional on deletes
Depending on the file type and its processing state, we may expect
nfo:FileDataObject and/or nie:InformationElement data in the content-
specific graphs.

The delete operations should account for those not being yet added,
or simply missing due to the file type. This fixes delete operations
after the nfo:FileDataObject/nie:InformationElement split.

This required a minor tweak on the corresponding test, so the fake
data we insert resembles more the current data layout as extracted by
the miner.