libtracker-miners-common: Add error storage infrastructure
Instead of logging things on stderr and asking users to look for it,
we are going for a coredumpctl-alike approach, where we persistently
log the errors in a way that can be quickly looked up and inspected
by the user.

This adds the lowlevel bits, so it is possible to store these errors.