tracker-extract: Avoid shutdown timeout if running from a terminal
Not strictly an user-facing CLI improvement, but this has been bugging
me forever. Make tracker-extract whether it's running on a tty, and avoid
the shutdown timeout in that case.

This is useful for debugging, as you can fire tracker-extract up, and
have it wait for triggered events, instead of timing everything perfectly.