tracker-extract: Set up watch on miner instance
And ensure we quit ASAP if that DBus name is gone. It doesn't matter
if we are running with a domain ontology or not, we can and should
quit tracker-extract on the same situation either way.

In domain ontologies, tracker-miner-fs still looks on the domain
DBus owner, so orderly shutdown will just be triggered indirectly.

OTOH, this allows us to indirectly shutdown everything by just pulling
the tracker-miner-fs process off.