tracker: Check recursive/single paths on "tracker3 index --add"
So we can't doubly add a directory to both settings.

It is worth noting here that there are other means to end up with
double folders (e.g. direct dconf manipulation), and tracker-miner-fs
copes with the situation (effectively, a folder that is both
non-recursive and recursive ends up being recursive). This is more
of an UI issue.