Log a g_message() on startup
The Tracker daemons used to be very verbose. We cleaned a lot of this up
behind TRACKER_DEBUG= but we went too far. I now find myself wondering
if logging to the journal is working. This commit guarantees that we
output at least one log message on startup to confirm that logging is
working.