tracker-miner-fs: Handle moving nfo:FileDataObject information in data graphs
If a file/dir is renamed, we handled both the nie:InformationElement in the
data graphs, and the nfo:FileDataObject in the filesystem graph. This got
broken when we started adding the nfo:FileDataObject shim information in the
data graphs.

Handle this, and split/comment the other queries so the role of each one is
clearer.

Fixes: https://gitlab.gnome.org/GNOME/tracker-miners/-/issues/136