tracker-extract: Insert SPARQL to ignore files that cause SPARQL errors
Bubble the error up from TrackerDecorator, and handle it like the other
failures in tracker-extract, so the file is reported, but doesn't sit
there waiting to be picked up for extraction again.

Now that we have "tracker3 status" and "tracker3 status $pattern", we
can patiently wait for user bug reports without melting the poles.