libtracker-miner: Handle move events from ineligible sources deeper
At the TrackerFileNotifier level, we just check that the file is in
the TrackerFileSystem, even though it's not guaranteed that only eligible
files will get there. Later on, we have checks at the TrackerMinerFS
level to handle these move events we let through as if the destination file
were new. We can let TrackerFileNotifier issue ::file-created instead of
::file-moved then.

If we trust the event flow, we should not get move events for files that
we didn't now about earlier (either previously from the database, or recently
from events that were handled right before this one), so the checks at the
TrackerMinerFS level become moot.