libtracker-miner: Avoid stat() on item_move()
If we trust the event flow, we shall receive a delete/move event for
that file in the future, so avoid handling it as a deletion here.

Besides the "peek into the future" layering break, we just used this
to see whether the file is a directory or not, we already have this
info from TrackerFileNotifier, so forward it.