libtracker-miner: Drop tracker_file_notifier_ensure_parents()
This is supposed to happen with MONITOR | !CHECK_TIME. Luckily or
not, !CHECK_MTIME is currently broken (doubly! in handling, and in
setting it), anyhow this flag shouldn't mean "ignore newly found
files/directories", when fixed in future commits.

Drop this, we should really issue ::file-created from the regular
paths.