libtracker-miner: Add missing ref to SPARQL buffer task
On error paths, we let tracker_sparql_buffer_push_finish() still return
the underlying TrackerTask, together with the error, so we can detach
it properly.

However, in this situation we are not adding an extra task ref, while
the callback expects to own a reference. This causes refcounting issues
and eventually invalid reads.

Fixes: https://gitlab.gnome.org/GNOME/tracker-miners/-/issues/130