libtracker-miner: Pass GFileInfo via TrackerFileNotifier::created/updated
We are duplicating work, we have the TrackerCrawler inspecting every folder
and file, just to queue changed files, and then we go through these again
querying info individually.

This can be streamlined by pushing the GFileInfos from the TrackerCrawler
instead, so we don't need to query that info again. This commit just
does one step forward, letting this information available outside the
TrackerFileNotifier.