libtracker-miner: Query items after commit on TrackerDecorator::finished
When the decorator finishes, we commit all pending tasks, and query again
just in case there's missed ones. If we don't wait for the former to do the
latter, we might be counting tasks that are actually completed.

We will check again after commit, so just don't do this here.