tracker-miner-fs: Repurpose crawling-interval setting
This setting (in days) was used to check whether tracker-miner-fs
would crawl on startup, after checking a timestamp file. However, we
are making crawling good and fast, and we don't like outdated databases.

Make this setting set up a timeout for long-running sessions instead,
and always perform crawling on startup.