build: Add -Dseccomp setting, default on
We need to keep seccomp disabled to content coverage reports, but
we don't want it to be easy to have it disabled. Make it a build
time option that we use in our CI, with a big warning about it
being malign.

Users: don't use -Dseccomp=false builds, even if it seems like it
fixes something.
Distributors: don't let other people use -Dseccomp=false builds.