libtracker-miner: Proceed to next folder in root after errors
When getting error on a folder, we just want to grind the index
root indexing to a halt when interrupted. Jumping directly to the
next root is a weird middle ground in that it goes ahead and
notifies as if finished, but then proceeds checking the missing
folders.

Let it maybe continue to the next folder in the root (unless
interrupted) so the whole index root is indexed (minus the failing
folder).