libtracker-miner: Fix TrackerMinerFS timer initialization
There was a path mistakenly re-starting the timer based on extraction
timer conditionals. This resets the timer every time, so we were
forgetting some time spent on TrackerFileNotifier andskewing the result.

While at it, try to be a bit more truthful, and account the time since
the miner is started (Not accounting: SPARQL connection initialization
time and startup timeout) after miner initialization.