tracker-miner-fs: Interpret GFileInfo uint64 times as time_t
This is a time_t underneath, and forcibly interpreting it as unsigned
will break with negative times on platforms that internally define that
type as signed (Linux and the GNU C library between them).

Deal with it as a time_t on our side, and let up to the underlying
implementation the interpretation of negative timestamps. This relies
on undocumented implementation details and is thus a hack.

Fixes: https://gitlab.gnome.org/GNOME/tracker-miners/-/issues/155

(cherry-picked from commit ebec93102c88db8378a5d23ab7e5441762a6bb26)