libtracker-miner: Make is_dir in TrackerFileNotifier::item-removed reliable
This argument, in the case of monitor events, comes right from the
TrackerMonitor. However this object just knows about the directories that
are being monitored, so the argument cannot be entirely trusted.

So use a query as a fallback, the database will know better the older
file type for the given URI. This will enable us to make optimizations
based on the file type without fear of missed cases.