tracker-miner-fs: Speed up delete of individual files
Querying the graphs that contain some data for the given file may be a bit
expensive, perhaps not that much noticeable during individual deletes, but
gets to show with many updates batched together.

It is sensibly faster to provide all known graphs to the query, and let
the individual DELETE operations do nothing in those graphs that don't have
any data for the file.

Fixes: https://gitlab.gnome.org/GNOME/tracker/-/issues/182