tracker-miner-fs: Use g_file_info_get_modification_date_time()
The interpretation of the uint64 G_FILE_ATTRIBUTE_TIME_MODIFIED value
with mtime < 0 is fairly undefined. We interpret it literally as an
uint64_t, but (signed) time_t is practically simply casted to it
internally in gio. This makes negative dates seem far far in the
future.

Use the GDateTime helper instead, this will a) leave the mtime
interpretation (bugs included) up to GLib. And b) will implicitly
make mtimes within the expected range, given Tracker and GDateTime
limits match.

Fixes: https://gitlab.gnome.org/GNOME/tracker-miners/-/issues/155