tracker-extract: Block gstreamer's nvcodec decoders
According to Fedora retrace service, we do seem to get a fair amount of
crashes in libcuda.so. E.g.:
https://retrace.fedoraproject.org/faf/problems/bthash/?bth=58896f17c609309b94b465e43ab9310821af02d6

This most likely comes from gstreamer, notably the nvcodec plugin
in gst-plugins-bad. Add all the decoders added by that plugin to the
blocklist, to ensure it's never pulled from the tracker-extract-3
process.