tests: Add timer after consecutive file operations
In tracker-miner-fs-test we are testing the simultaneous handling
of several consecutive file operations. We however don't get any
guarantees that these all will be handled atomically by the time
we iterate the main context to process the events.

This may result in split operations, that stop the miner in between,
thus ends our testcases without the expected results.

In these situations, wait for all events with a timeout, instead
of looking for TrackerMiner being finished. We ensure slightly
better that inotify/glib and our own TrackerMonitor thread did
all catch up to all the changes within the fixture_run_timed()
calls.

Fixes: https://gitlab.gnome.org/GNOME/tracker-miners/-/issues/177