libtracker-miner: Disconnect from the TrackerMonitor on finalize
When finalizing the TrackerFileNotifier, we may presumably not hold
the last reference to the TrackerMonitor, there may be events in
flight not yet handled by the main thread, and these might be still
created until the moments that TrackerMonitor itself is being finalized.
These events hold each a reference to the TrackerMonitor that might
make it live longer than it is expected here.

Ensure to hint shutdown to the TrackerMonitor, and disconnect from
any pending signals before finalizing it. If there are pending events
these will be emitted into the void and freed, eventually tearing down
the TrackerMonitor.

Avoids possible crashes after finalizing the TrackerFileNotifier/Monitor,
as these signals would poke invalid memory, or objects being finalized,
apparently can be occasionally seen running tracker-miner-fs-test to
exhaustion.