libtracker-miner: Fix possible double DELETED emission on directories
We may receive DELETED events in directories from 2 directory monitors:
- The directory monitor for this directory itself (via IN_DELETE_SELF)
- The directory monitor of the parent directory (via regular IN_DELETE
  on a child).

We typically will get the parent directory event first, cancel the
directory monitor for the deleted directory, and this way avoid receiving
the second event.

Less typically, we would receive the event from the directory itself
first, cancel that file monitor, but still receive the second DELETED
event from the parent folder (which was not cancelled). In these cases,
we'd usually still coalesce the events anyway, since the DELETED event
emission does not happen immediately (and the DELETED event would drop
the cached events).

But, in even stranger situations (I presume load related, since this
has just been seen in CI an build machines), these 2 DELETED events
happen separately enough that the monitor thread already let the first
DELETED event entirely through, so no coalescing happens, and we actually
emit DELETED on the folder twice. This caused test failures in
tracker-file-notifier-test.

To fix this, check explicitly on DELETED events the folder's directory
monitor and whether it's already cancelled, before maybe emitting a
second event. This still does account for the cases where there is only
one directory monitor on the file (e.g. parent is monitored, but we run
out of monitors with the directory itself).

Fixes: https://gitlab.gnome.org/GNOME/tracker-miners/-/issues/174