libtracker-miner: Fix TrackerMinerFS event coalescing
This infrastructure relied on GFiles being unique-ified by the
TrackerFileSystem. Since we remove that piece of caching, GFiles
for the same path can actually be different instances at this
level, so we should account for this.

Keep a hash table using GFile equal/hash functions, so we can
still have this event coalescing in place.

The way this was missed feels rather disconnected, when a directory
and a child file are created in a quick sucession, it may happen
that the TrackerFileNotifier did add a monitor for the directory,
started crawling it, and found this new file on both paths. If
this happens, the 2 CREATED events will be propagated upwards.

Since crawling+monitoring is inherently racy (and inverting the
order may actually miss events), we have to deal with this double
emission since we cannot handle it within TrackerFileNotifier.
This was usually handled at the TrackerMinerFS event queue level
by this event coalescing.

This situation was noticed during investigation of
https://gitlab.gnome.org/GNOME/tracker-miners/-/issues/179, there
are additional issues in the Tracker library leading to the warning,
but this is first thing that starts going wrong there, and this
fix is sufficient to avoid the issue entirely.

Fixes: https://gitlab.gnome.org/GNOME/tracker-miners/-/issues/179