tests: Await for all resources being deleted
In miner-basic's 09_deletion_directory test, we have a directory
containing 2 text files, and test that removing the directory ends
up with these 2 files removed.

However the test infrastructure only allows awaiting for one
resource at a time, so we pick one of those text files to monitor
for removal while removing the directory, and call the whole operation
done when that did happen.

However the ordering and timing of events is racy here, we could
receive the expected event and proceed while the other file was not
deleted yet from the database. When that happens, this is seen as:

  AssertionError: 2 != 1

In that testcase, since we are expecting to see one text document
from a different folder, but we additionally see the not-yet-deleted
file. To fix this, ensure (in a more lenient way) that the other
text file is also deleted in a reasonable time.

Fixes: https://gitlab.gnome.org/GNOME/tracker-miners/-/issues/185